Feat : "회원가입 예외처리"

+++ src/main/java/com/finalProject/Back/repository/OwnerMapper.java
@@ -12,5 +12,5 @@ public interface OwnerMapper {
     List<RespGetUserDto> getUsers();
     List<RespGetOwnerDto> getOwners();
     Long deleteUser(Long id);
-    int deleteCafe(int id);
+    Long deleteCafe(Long id);
 }
